using uint8_t** as pbm now

+++ lib/pbm.h
@@ -1,5 +1,5 @@
 #include <stdint.h>
 
-uint64_t** pbm_init(void);
-int pbm_check(uint64_t **b, uint32_t v);
-void pbm_set(uint64_t **b, uint32_t v);
+uint8_t** pbm_init(void);
+int pbm_check(uint8_t **b, uint32_t v);
+void pbm_set(uint8_t **b, uint32_t v);
